Some synchronization for objects. Please review the code.
The code is based on ped and unoccupied vehicle sync. Some code was copied, probably I forgot to update something.

Sync rate and distance copied from ped sync. I'm not sure that these values are good for objects.

Static and unbreakable objects are not synced.

Synced values are position, rotation and health. Object breaking is not synced in this commit, but I'm going to add it later.