[Eir]: Phase 4 - New Features - reverse engineered the remainder of the engine caching logic - reverse engineered the remainder of the IPL container loading (still WIP; gotta wire things together)
This is one of those hardcore reven commits. It is experimental for now, too. Now the loading of buildings by the GTA:SA engine is open for everyone to see. It is pretty weird too, convoluted with weird bugfixes by R* and array links that reach multiple call stack layers back.

Two important factors have been established:
- I can now look at how buildings are properly created, so createBuilding can be worked on.
- I can reimplement the lost CINFO.BIN feature to speed up GTA:SA engine startup

As this is just a temporary junction, lets go ahead.